Move all ChangeLog elements into a separate directory
